+++ libprint.c
@@ -6,7 +6,7 @@
 /*   By: ocarlos- <ocarlos-@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/04/30 21:45:56 by ocarlos-          #+#    #+#             */
-/*   Updated: 2020/07/21 16:07:37 by ocarlos-         ###   ########.fr       */
+/*   Updated: 2020/07/21 16:26:20 by ocarlos-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int		ft_pchr(char c, t_prtform *form)
 	return (1);
 }
 
-int		ft_phex(unsigned long int i, t_prtform *form, int swt) //***mudar aqui
+int		ft_phex(unsigned long long int i, t_prtform *form, int swt) //***mudar aqui
 {
 	char	*s;
 	char	*temp;
@@ -90,7 +90,7 @@ int		ft_pstr(char *s, t_prtform *form)
 	return (form->siz);
 }
 
-int		ft_puin(unsigned long int i, t_prtform *form) //***mudar aqui
+int		ft_puin(unsigned long long int i, t_prtform *form) //***mudar aqui
 {
 	char	*s;
